+++ csv_database.csv
@@ -3,4 +3,5 @@
 2,testusername2,testemail@gmail.com,scrypt:32768:8:1$Ay2yz91pih18kPMX$90aa3853f5c6f0ab50bd63fd41916e70fa7790f2ad2ce31929c73d5bed044e80b26f14ef93ca8bd96fd54c0d24c037d68d44423d2496c5bb8cc8e138a2b67bac
 3,testusername3,,scrypt:32768:8:1$JS0lzKNvXHuCy9Mm$c825f082245e8e209fedfbc6571a9604e6fa34d8dde7d765fcb971cfdf484cb171ab33c54682b40e04760175ef1f2fded56ce2876b5216b258bac56395ae2efc
 4,testusername4,user@yahoo.com,scrypt:32768:8:1$3Dz5lFEBVn2gytre$5110a2f53e1a04d0523e6e2b8f502e16e0ee29a76f2da314313edf8014f19f8c9d063e98ba30a22e3693a90547fb5c0e9110b6bdb6e3bb07b52112e20da801f4
-5,testusername14,,scrypt:32768:8:1$GpEou1Gwy7ICHcwY$cdbf7c6ec2f031633869b6e3f88a055ed2c652d2381dd1ca0e4d426120bd5a3933441fa3fbd86344e2486f36d0185bccb95400b17bf873db7d98200f3b5fbae7
+5,testusername14,,scrypt:32768:8:1$GpEou1Gwy7ICHcwY$cdbf7c6ec2f031633869b6e3f88a055ed2c652d2381dd1ca0e4d426120bd5a3933441fa3fbd86344e2486f36d0185bccb95400b17bf873db7d98200f3b5fbae76,integration_test,,scrypt:32768:8:1$pybMYi9CksPxKEfN$f80640bc1b8617922f1eae44ab46743468afafc86a361bf1dfb05159c121c69b600a4012b79b541925c32359523253e847b7c14175faf13c1045ea60f2dc167f
+6,integration_test,,scrypt:32768:8:1$h2JOEHjBeH1oelCW$1e1d1e0c78e97236135317b1613560257059da8c35c6e8ca2bf491952276d00b46284bbe3676b3e34c7c4c73ffd8403656984f1c67c2ec57adb80cc085016c45
