+++ templates/index.html
@@ -7,5 +7,6 @@
 </head>
 <body>
 <p>Welcome To RoomMate</p>
+<button id="createAccount" type = "button" onclick= "window.location.href='/create_account' ; " class="button">Create Account</button>
 </body>
 </html>
